mnist/requirements.txt
pymongo==4.3.3
tensorflow-macos
numpy==1.23.5
pandas==1.5.2
matplotlib==3.6.2